+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Prevent to display a deleted video
 - Disable video speed menu for live video
+- Manage jitsi recording interruption
 
 ### Removed